Andere Datentypen

+++ blatt2/src/Graph.h
@@ -10,9 +10,9 @@
 // Define types for ids and values
 typedef unsigned int node_t;
 typedef unsigned int edge_t;
-typedef int flow_t;
-typedef int supply_t;
-typedef int cost_t;
+typedef long long int flow_t;
+typedef long long int supply_t;
+typedef long long int cost_t;
 typedef long long int pot_t;
 
 class Graph {
